adds the protection to check if the updating or deleting request is coming exclusively from the owner of that object/account

+++ src/main/java/presto/com/FoodDeliveryAPI/service/AuthService.java
@@ -28,6 +28,9 @@ public class AuthService implements UserDetailsService {
     StoreRepository storeRepository;
 
     @Autowired
+    UtilityService utilityService;
+
+    @Autowired
     private TokenService tokenService;
 
     @Override
@@ -61,6 +64,8 @@ public class AuthService implements UserDetailsService {
         if(!credentialsRepository.existsById(id)){
             throw new EntityNotFoundException("Nenhuma conta associada a este Id foi encontrada.");
         }
+
+        utilityService.checkPermission(id);
 
         credentialsRepository.deleteById(id);
     }
